requirements.txt
fastapi
uvicorn
ultralytics
onnxruntime
numpy
Pillow
python-multipart
tensorflow